+++ src/nonempty_list.mli
@@ -3,7 +3,8 @@ open Base
 (** A ['a t] represents a non-empty list, as evidenced by the fact that there is no [[]]
     variant. The sexp representation is as a regular list (i.e., the same as the
     [Stable.V3] module below). *)
-type 'a t = ( :: ) of 'a * 'a list [@@deriving compare, equal, sexp, sexp_grammar, hash]
+type 'a t = ( :: ) of 'a * 'a list
+[@@deriving compare, equal, sexp, sexp_grammar, hash, quickcheck]
 
 include Comparator.Derived with type 'a t := 'a t
 include Container.S1 with type 'a t := 'a t
@@ -29,6 +30,8 @@ val unzip : ('a * 'b) t -> 'a t * 'b t
 val zip : 'a t -> 'b t -> ('a * 'b) t List.Or_unequal_lengths.t
 val zip_exn : 'a t -> 'b t -> ('a * 'b) t
 val mapi : 'a t -> f:(int -> 'a -> 'b) -> 'b t
+val map2 : 'a t -> 'b t -> f:('a -> 'b -> 'c) -> 'c t List.Or_unequal_lengths.t
+val map2_exn : 'a t -> 'b t -> f:('a -> 'b -> 'c) -> 'c t
 val filter : 'a t -> f:('a -> bool) -> 'a list
 val filteri : 'a t -> f:(int -> 'a -> bool) -> 'a list
 val filter_map : 'a t -> f:('a -> 'b option) -> 'b list
@@ -40,7 +43,8 @@ val drop_last : 'a t -> 'a list
 val to_sequence : 'a t -> 'a Sequence.t
 val sort : 'a t -> compare:('a -> 'a -> int) -> 'a t
 val stable_sort : 'a t -> compare:('a -> 'a -> int) -> 'a t
-val dedup_and_sort : compare:('a -> 'a -> int) -> 'a t -> 'a t
+val dedup_and_sort : 'a t -> compare:('a -> 'a -> int) -> 'a t
+val permute : ?random_state:Random.State.t -> 'a t -> 'a t
 val iteri : 'a t -> f:(int -> 'a -> unit) -> unit
 val cartesian_product : 'a t -> 'b t -> ('a * 'b) t
 val fold_nonempty : 'a t -> init:('a -> 'acc) -> f:('acc -> 'a -> 'acc) -> 'acc
@@ -70,6 +74,14 @@ val map_of_alist_multi
 val map_of_sequence_multi
   :  ('k * 'v) Sequence.t
   -> comparator:('k, 'cmp) Comparator.Module.t
+  -> ('k, 'v t, 'cmp) Map.t
+
+(** Like [Map.of_list_with_key_multi], but comes with a guarantee that the range of the
+    returned map is all nonempty lists. *)
+val map_of_list_with_key_multi
+  :  'v list
+  -> comparator:('k, 'cmp) Comparator.Module.t
+  -> get_key:('v -> 'k)
   -> ('k, 'v t, 'cmp) Map.t
 
 (** Like [Result.combine_errors] but for non-empty lists *)